# main/docs/swagger.yaml
basePath: /
definitions:
...
  dto.LocaleResponse:
    properties:
      en:
        description: 英语
        type: string
      ja:
        description: 日语
        type: string
      ko:
        description: 韩语
        type: string
      my:
        description: 缅甸语
        type: string
      th:
        description: 泰语
        type: string
      tr:
        description: 土耳其语
        type: string
      zh:
        description: 中文
        type: string
      zhtw:
        description: 粤语
        type: string
    type: object
...
  dto.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  req.OrderCancelReq:
    properties:
      cancel_reason:
        description: 取消原因
        type: string
      notNeedPassword:
        description: 不需要验证高级密码。仅用后端使用
        type: boolean
      password:
        description: 高级密码 后台开启的时候才传
        type: string
      sale_bill_uuid:
        description: 销售账单UUID
        type: integer
    type: object
...
  req.OrderDeleteReq:
    properties:
      sale_bill_uuid:
        description: 销售账单UUID
        type: integer
      sale_order_uuid:
        description: 销售订单UUID 传0的时候默认删除主单以及所有子单，不然只删除子单
        type: integer
    type: object
...
  req.OrderPrintReq:
    properties:
      pay_method_uuid:
        description: 支付方式UUID, 可选 (打印码时用)
        type: integer
      print_lang:
        description: 打印语言, 可选
        type: string
      sale_bill_uuid:
        description: 销售账单UUID, 必填
        type: integer
      sale_order_uuid:
        description: 销售订单UUID, 必填
        type: integer
    type: object
...
  req.OrderReturnProduct:
    properties:
      num:
        description: 退款数量
        type: integer
      sale_order_product_uuid:
        description: 销售订单商品UUID
        type: integer
    type: object
  req.OrderReturnReq:
    properties:
      products:
        description: 退款商品UUID列表. 如果为空，则退款所有商品,即整单退款
        items:
          $ref: '#/definitions/req.OrderReturnProduct'
        type: array
      sale_bill_uuid:
        description: 销售账单UUID
        type: integer
      sale_order_uuid:
        description: 销售订单UUID。退款都是针对子单进行退款
        type: integer
    type: object
  req.OrderReverseSettleReq:
    properties:
      desk_uuid:
        description: 桌台UUID. 仅桌台订单才填该字段
        type: integer
      hide_order:
        description: 是否挂单. 仅点餐订单才填该字段
        type: boolean
      sale_bill_uuid:
        description: 销售账单UUID
        type: integer
    type: object
...
  resp.BillLists:
    properties:
      bill_type:
        description: "订单类型\t0:桌台订单 1:点餐订单"
        type: integer
      consumer_uuids:
        description: 会员id
        type: string
      extra:
        allOf:
        - $ref: '#/definitions/resp.BillListsExtra'
        description: 通过当前数据控制按钮是否显示
      finish_time:
        description: 完成时间（支付时间）（时间戳）
        type: integer
      is_split:
        description: "是否拆单\tfalse:否 true:是"
        type: boolean
      order_amount:
        description: 订单总金额
        type: number
      order_no:
        description: 订单编号
        type: string
      pay_type_name:
        description: 支付类型名称
        type: string
      payment_amount:
        description: 支付金额
        type: number
      sale_bill_uuid:
        description: 销售账单UUID
        type: integer
      sale_order_uuid:
        description: 销售订单UUID,第一个销售订单的uuid
        type: integer
      sale_orders:
        description: 订单列表
        items:
          $ref: '#/definitions/resp.BillListsOrder'
        type: array
      serial_no:
        description: 桌位编号 (点餐流水号)
        type: string
      status:
        description: 订单状态 订单状态, 0-待付款、1-已完成、2-已取消
        type: integer
    type: object
  resp.BillListsExtra:
    properties:
      is_cell_cancel:
        description: 是否可取消
        type: boolean
      is_cell_delete:
        description: 是否可删除
        type: boolean
      is_cell_invoice:
        description: 是否可打印发票
        type: boolean
      is_cell_print:
        description: 是否可打印小票
        type: boolean
      is_cell_refund:
        description: 是否可退款
        type: boolean
      is_cell_reverse_settle:
        description: 是否可反结账
        type: boolean
    type: object
  resp.BillListsOrder:
    properties:
      bill_type:
        description: "订单类型\t0:桌台订单 1:点餐订单"
        type: integer
      consumer_uuids:
        description: 会员id
        type: string
      extra:
        $ref: '#/definitions/resp.BillListsExtra'
      finish_time:
        description: 完成时间（支付时间）（时间戳）
        type: integer
      order_amount:
        description: 订单总金额
        type: number
      order_no:
        description: 订单编号
        type: string
      pay_type_name:
        description: 支付类型名称
        type: string
      payment_amount:
        description: 支付金额
        type: number
      sale_bill_uuid:
        description: 销售订单UUID
        type: integer
      sale_order_uuid:
        description: 销售订单UUID
        type: integer
      serial_no:
        description: 桌位编号 (点餐流水号)
        type: string
      status:
        description: 订单状态 订单状态, 0-待付款、1-已完成、2-已取消
        type: integer
    type: object
...
  resp.OrderInfo:
    properties:
      bill_type:
        description: "订单类型\t0:桌台订单 1:点餐订单"
        type: integer
      dining_method:
        description: 用餐方式,0-堂食 1-打包
        type: integer
      finish_time:
        description: 完成时间（支付时间）（时间戳）
        type: integer
      free_reason:
        description: 免单原因
        type: string
      is_free:
        description: 是否免单
        type: boolean
      member_name:
        description: 会员名称
        type: string
      member_uuid:
        description: 会员名称
        type: integer
      order_amount:
        description: 订单总金额
        type: number
      order_no:
        description: 订单编号
        type: string
      pay_type_name:
        description: 支付类型名称
        type: string
      payment_amount:
        description: 支付金额
        type: number
      products:
        description: 产品列表
        items:
          $ref: '#/definitions/resp.OrderProduct'
        type: array
      refund_amount:
        description: 退款金额
        type: number
      sale_order_uuid:
        description: 销售订单UUID
        type: integer
      serial_no:
        description: 桌位编号 (点餐流水号)
        type: string
      status:
        description: 订单状态 订单状态, 0-待付款、1-已完成、2-已取消
        type: integer
    type: object
  resp.OrderInfoPayTypes:
    properties:
      code:
        description: 支付Code
        type: string
      currency_unit:
        description: 货币单位
        type: string
      payment_amount:
        description: 支付金额
        type: number
      payment_type_name:
        description: 支付类型名称
        type: string
      sale_bill_uuid:
        description: 支付单UUID
        type: integer
      source:
        description: 支付状态, 来源 0-系统 1-手动 2-LianLianPay
        type: integer
      source_text:
        description: 支付状态, 来源 0-系统 1-手动 2-LianLianPay
        type: string
      status:
        description: 支付状态, 0-未支付 1-已支付 2-已退款
        type: integer
    type: object
  resp.OrderInfos:
    properties:
      bill_type:
        description: "订单类型\t0:桌台订单 1:点餐订单"
        type: integer
      buffet_names:
        description: 自助餐名称
        type: string
      cancel_reason:
        description: 取消原因
        type: string
      cashier_name:
        description: 收银员名称
        type: string
      create_time:
        description: 创建时间
        type: integer
      dining_method:
        description: 用餐方式,0-堂食(店内就餐) 1-打包
        type: integer
      finish_time:
        description: 完成时间（支付时间）（时间戳）
        type: integer
      is_buffet:
        description: "是否自助餐\tfalse:否 true:是"
        type: boolean
      is_split:
        description: "是否拆单\tfalse:否 true:是"
        type: boolean
      member_names:
        description: 会员名称
        type: string
      member_uuids:
        description: 会员名称
        type: string
      order_amount:
        description: 订单总金额
        type: number
      order_no:
        description: 订单编号
        type: string
      pay_types:
        description: 支付类型
        items:
          $ref: '#/definitions/resp.OrderInfoPayTypes'
        type: array
      payment_amount:
        description: 支付金额
        type: number
      refund_amount:
        description: 退款金额
        type: number
      remark:
        description: 备注
        type: string
      sale_bill_uuid:
        description: 销售账单UUID
        type: integer
      sale_orders:
        description: 订单列表
        items:
          $ref: '#/definitions/resp.OrderInfo'
        type: array
      serial_no:
        description: 桌位编号 (点餐流水号)
        type: string
      status:
        description: 订单状态 订单状态, 0-待付款、1-已完成、2-已取消
        type: integer
    type: object
  resp.OrderInfosResp:
    properties:
      detail:
        allOf:
        - $ref: '#/definitions/resp.OrderInfos'
        description: 明细
      extra:
        allOf:
        - $ref: '#/definitions/resp.BillListsExtra'
        description: 通过当前数据控制按钮是否显示
      operation_log:
        description: 操作日志
        properties:
          list:
            items:
              $ref: '#/definitions/resp.OrderOperationLog'
            type: array
        type: object
    type: object
  resp.OrderListMeta:
    properties:
      cancel_num:
        description: 已取消数量
        type: integer
      complete_num:
        description: 已完成数量
        type: integer
      page_no:
        description: 当前页码
        type: integer
      page_size:
        description: 每页大小
        type: integer
      total:
        description: 总数
        type: integer
      unpaid_num:
        description: 待付款数量
        type: integer
    type: object
  resp.OrderListPaginationResp:
    properties:
      list:
        description: 订单列表
        items:
          $ref: '#/definitions/resp.BillLists'
        type: array
      meta:
        allOf:
        - $ref: '#/definitions/resp.OrderListMeta'
        description: Meta信息
    type: object
  resp.OrderOperationLog:
    properties:
      action:
        description: 操作行为
        type: string
      create_time:
        description: 创建时间(时间戳)
        type: integer
      data:
        description: 消息数据
      remark:
        description: 备注
        type: string
      sale_bill_uuid:
        description: 销售账单ID
        type: integer
      sale_order_uuid:
        description: 销售订单ID
        type: integer
      source:
        description: 操作来源 cashier-收银 assistant-助手 shop-商家后台
        type: string
      uuid:
        description: 账单操作记录ID
        type: integer
    type: object
  resp.OrderProduct:
    properties:
      attributes:
        description: 规格属性加料
        type: string
      flavor_name:
        description: 口味名称
        type: string
      gift_reason:
        description: 赠品原因
        type: string
      image_url:
        description: 图片地址
        type: string
      is_gift:
        description: 是否赠品, false-否 true-是
        type: boolean
      locale_name:
        allOf:
        - $ref: '#/definitions/dto.LocaleResponse'
        description: 产品名称
      num:
        description: 数量
        type: integer
      price:
        description: 最终单价(单商品，会员、会员卡和优惠折扣后，折后价)。销售价*折扣率'
        type: number
      refund_amount:
        description: 退款金额
        type: number
      refund_reason:
        description: 退菜原因
        type: string
      remark:
        description: 备注
        type: string
      sale_price:
        description: 销售价（单商品，折前价）,当自定义价格时，销售价=自定义价格,否则销售价=原始单价
        type: number
      status:
        description: 状态, 0-正常 1-退菜
        type: integer
      tax_rate:
        description: 税率,单位%.下单时单税率,结账时再重新核算
        type: number
      total_price:
        description: 应收金额(单商品)=最终单价+服务费+总税费
        type: number
      total_sale_price:
        description: 销售价（单商品，折前价）,当自定义价格时，销售价=自定义价格,否则销售价=原始单价'
        type: number
      uuid:
        description: 销售订单商品ID
        type: integer
    type: object
  resp.OrderReturnInfoResp:
    properties:
      can_return_amount:
        description: 可退款金额. 可退款金额=订单最终应收金额-已退款金额
        type: number
      payment_records:
        description: 支付记录
        items:
          $ref: '#/definitions/resp.OrderReturnPaymentRecord'
        type: array
      products:
        description: 商品列表
        items:
          $ref: '#/definitions/resp.OrderReturnProduct'
        type: array
    type: object
  resp.OrderReturnPaymentRecord:
    properties:
      can_return_amount:
        description: 可退款金额。剩余可退xx. 可退款金额=支付金额-已退款金额
        type: number
      currency_unit:
        description: 货币单位
        type: string
      payment_amount:
        description: 支付金额
        type: number
      payment_method_name:
        description: 支付类型名称
        type: string
      payment_order_uuid:
        description: 支付单UUID
        type: integer
    type: object
  resp.OrderReturnProduct:
    properties:
      can_return_amount:
        description: 可退款金额. 可退款金额=订单商品数量*最终单价
        type: number
      currency_unit:
        description: 货币单位
        type: string
      locale_attribute_name:
        allOf:
        - $ref: '#/definitions/dto.LocaleResponse'
        description: 属性名称
      locale_name:
        allOf:
        - $ref: '#/definitions/dto.LocaleResponse'
        description: 产品名称
      num:
        description: 数量。可退货数量=订单商品数量-已退货数量
        type: integer
      price:
        description: 最终单价(单商品，会员、会员卡和优惠折扣后，折后价)。
        type: number
      sale_order_product_uuid:
        description: 销售订单商品ID
        type: integer
    type: object
  resp.OrderReverseSettleDesk:
    properties:
      serial_no:
        description: 桌台编号
        type: string
      uuid:
        description: 桌台UUID
        type: integer
    type: object
  resp.OrderReverseSettleDeskList:
    properties:
      list:
        description: 桌台编号
        items:
          $ref: '#/definitions/resp.OrderReverseSettleDesk'
        type: array
      origin_desk_available:
        description: 原桌台是否空闲
        type: boolean
    type: object
  resp.OrderReverseSettleInfoResp:
    properties:
      desks:
        allOf:
        - $ref: '#/definitions/resp.OrderReverseSettleDeskList'
        description: 可用桌台列表。仅桌台订单才有该字段
      has_instant_order:
        description: 是否存在即时订单。仅点餐订单才有该字段
        type: boolean
      order_amount:
        description: 订单总金额
        type: number
      pay_methods:
        description: 支付方式名称列表
        items:
          type: string
        type: array
      payment_amount:
        description: 支付金额
        type: number
      sale_bill_no:
        description: 销售账单编号
        type: string
      sale_bill_type:
        description: 销售账单类型.0-桌台订单、1-点餐订单
        type: integer
      sale_bill_uuid:
        description: 销售账单UUID
        type: integer
    type: object
...
  resp.PrinterData:
    properties:
      copies:
        description: 打印机.份数
        type: integer
      data:
        description: 打印数据
        type: string
      print_method:
        description: 打印方式 1文本打印, 2图片打印'
        type: integer
      printer_config:
        description: 打印机.配置
        type: string
      printer_type:
        description: 打印机.类型 打印机类型 (SUNMI_LAN:商米打印机, SUNMI_CLOUD:商米打印机-云, XPRINTER_LAN:芯烨-有线
          , XPRINTER_WIFI:芯烨-WIFI , CASHIER:收银机自带打印机)
        type: string
      uuid:
        description: 打印日志Uuid
        type: integer
    type: object
...
host: 127.0.0.1:8081
info:
  contact: {}
  description: 点餐系统服务端接口文档
  title: ttpos-server-go API
  version: "1.0"
paths:
...
  /cashier/order/cancel:
    post:
      consumes:
      - application/json
      description: 取消订单
      parameters:
      - description: 详情参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/req.OrderCancelReq'
      produces:
      - application/json
      responses:
        "200":
          description: 取消订单成功
        "404":
          description: 未找到
      security:
      - JwtToken: []
      summary: 取消订单
      tags:
      - 收银端.订单
  /cashier/order/delete:
    delete:
      consumes:
      - application/json
      description: 删除订单
      parameters:
      - description: 详情参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/req.OrderDeleteReq'
      produces:
      - application/json
      responses:
        "200":
          description: 取消订单成功
        "404":
          description: 未找到
      security:
      - JwtToken: []
      summary: 删除订单
      tags:
      - 收银端.订单
  /cashier/order/info:
    get:
      consumes:
      - application/json
      description: 获取订单详情
      parameters:
      - description: 销售账单UUID
        in: query
        name: saleBillUuid
        type: integer
      - description: 销售订单UUID 当查看子订单信息的时候才需要传
        in: query
        name: saleOrderUuid
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 订单详情
          schema:
            $ref: '#/definitions/resp.OrderInfosResp'
        "404":
          description: 未找到
      security:
      - JwtToken: []
      summary: 获取订单详情
      tags:
      - 收银端.订单
  /cashier/order/is_cell_close:
    get:
      consumes:
      - application/json
      description: 判断订单是否可关闭
      parameters:
      - description: "桌台UUID\t   二选一, 销售账单UUID权重最大"
        in: query
        name: deskUuid
        type: integer
      - description: "销售账单UUID\t二选一，销售账单UUID权重最大"
        in: query
        name: saleBillUuid
        type: integer
      produces:
      - application/json
      responses:
        "404":
          description: 未找到
      security:
      - JwtToken: []
      summary: 判断订单是否可关闭
      tags:
      - 收银端.订单
  /cashier/order/list:
    get:
      consumes:
      - application/json
      description: 获取订单列表
      parameters:
      - description: 账单类型, -1=全都、 0=Desk桌台订单、1=OrderingFood点餐订单
        in: query
        name: billType
        type: integer
      - description: 日期类型 -1=全都、 0=今天、 1=昨天、 2=本周
        in: query
        name: dateType
        type: integer
      - description: 用餐方式,-1=全都、 0-堂食 1-打包
        in: query
        name: diningMethod
        type: integer
      - description: 启用开台时间 false-不启用，true-启用
        in: query
        name: enableCreateTime
        type: boolean
      - description: 启用支付时间 false-不启用，true-启用
        in: query
        name: enablePayTime
        type: boolean
      - description: 订单编号
        in: query
        name: orderNo
        type: string
      - description: 页码
        in: query
        minimum: 1
        name: page_no
        type: integer
      - description: 每页大小
        in: query
        maximum: 1000
        minimum: 1
        name: page_size
        type: integer
      - description: 查询结束时间戳
        in: query
        name: queryEndTime
        type: integer
      - description: 查询开始时间戳
        in: query
        name: queryStartTime
        type: integer
      - description: 账单状态, -1=全都、 0=待付款、1=已完成、2=已取消
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 订单列表
          schema:
            $ref: '#/definitions/resp.OrderListPaginationResp'
        "404":
          description: 未找到
      security:
      - JwtToken: []
      summary: 获取订单列表
      tags:
      - 收银端.订单
  /cashier/order/print:
    post:
      consumes:
      - application/json
      description: 打印
      parameters:
      - description: 打印
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/req.OrderPrintReq'
      produces:
      - application/json
      responses:
        "200":
          description: 打印数据
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.PrinterData'
              type: object
      security:
      - JwtToken: []
      summary: 打印
      tags:
      - 收银端.订单
  /cashier/order/return:
    get:
      consumes:
      - application/json
      description: 获取退款信息
      parameters:
      - description: 销售账单UUID
        in: query
        name: saleBillUuid
        type: integer
      - description: 销售订单UUID。退款都是针对子单进行退款
        in: query
        name: saleOrderUuid
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.OrderReturnInfoResp'
              type: object
        "404":
          description: 未找到
      security:
      - JwtToken: []
      summary: 获取退款信息
      tags:
      - 收银端.订单
    post:
      consumes:
      - application/json
      description: 退款订单
      parameters:
      - description: 详情参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/req.OrderReturnReq'
      produces:
      - application/json
      responses:
        "200":
          description: 退款订单成功
        "404":
          description: 未找到
      security:
      - JwtToken: []
      summary: 退款订单
      tags:
      - 收银端.订单
  /cashier/order/reverse_settle:
    get:
      consumes:
      - application/json
      description: 获取反结账弹窗信息
      parameters:
      - description: 销售账单UUID
        in: query
        name: saleBillUuid
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.OrderReverseSettleInfoResp'
              type: object
        "404":
          description: 未找到
      security:
      - JwtToken: []
      summary: 获取反结账弹窗信息
      tags:
      - 收银端.订单
    post:
      consumes:
      - application/json
      description: 反结账
      parameters:
      - description: 详情参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/req.OrderReverseSettleReq'
      produces:
      - application/json
      responses:
        "200":
          description: 反结账成功
        "404":
          description: 未找到
      security:
      - JwtToken: []
      summary: 反结账
      tags:
      - 收银端.订单
...
securityDefinitions:
  JwtToken:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"